working through test fixes

+++ vowpalwabbit/search.cc
@@ -2170,7 +2170,7 @@ void end_pass(search& sch)
     //reset search_trained_nb_policies in options_from_file so it is saved to regressor file later
     // TODO work out a better system to update state that will be saved in the model.
     all->options->replace("search_trained_nb_policies", std::to_string(priv.current_policy));
-    all->options->get_typed_option<size_t>("search_trained_nb_policies").value(static_cast<size_t>(priv.current_policy));
+    all->options->get_typed_option<uint32_t>("search_trained_nb_policies").value(priv.current_policy);
   }
 }
 
@@ -2191,7 +2191,7 @@ void end_examples(search& sch)
     // Dig out option and change it in case we already loaded a predictor which had a value stored for --search_trained_nb_policies
     auto val = (priv.passes_since_new_policy == 0) ? priv.current_policy : (priv.current_policy+1);
     all->options->replace("search_trained_nb_policies", std::to_string(val));
-    all->options->get_typed_option<size_t>("search_trained_nb_policies").value(val);
+    all->options->get_typed_option<uint32_t>("search_trained_nb_policies").value(val);
     // Dig out option and change it in case we already loaded a predictor which had a value stored for --search_total_nb_policies
     all->options->replace("search_total_nb_policies", std::to_string(priv.total_number_of_policies));
     all->options->get_typed_option<uint32_t>("search_total_nb_policies").value(priv.total_number_of_policies);
@@ -2437,7 +2437,7 @@ base_learner* setup(options_i& options, vw& all)
   std::string rollout_string = "mix_per_state";
   std::string rollin_string = "mix_per_state";
 
-  size_t search_trained_nb_policies;
+  uint32_t search_trained_nb_policies;
   std::string search_allowed_transitions;
 
   priv.A = 1;
@@ -2546,10 +2546,10 @@ base_learner* setup(options_i& options, vw& all)
     priv.current_policy--;
 
   all.options->replace("search_trained_nb_policies", std::to_string(priv.current_policy));
-  all.options->get_typed_option<size_t>("search_trained_nb_policies").value(static_cast<size_t>(priv.current_policy));
+  all.options->get_typed_option<uint32_t>("search_trained_nb_policies").value(priv.current_policy);
 
   all.options->replace("search_total_nb_policies", std::to_string(priv.total_number_of_policies));
-  all.options->get_typed_option<size_t>("search_total_nb_policies").value(static_cast<size_t>(priv.total_number_of_policies));
+  all.options->get_typed_option<uint32_t>("search_total_nb_policies").value(priv.total_number_of_policies);
 
   cdbg << "search current_policy = " << priv.current_policy << " total_number_of_policies = " << priv.total_number_of_policies << endl;
